修复告警1104

+++ frameworks/core/components_ng/pattern/text_field/content_controller.cpp
@@ -352,7 +352,7 @@ bool ContentController::FilterWithEvent(const std::string& filter, std::string& 
 
 void ContentController::erase(int32_t startIndex, int32_t length)
 {
-    if (startIndex < 0 || startIndex >= GetWideText().length()) {
+    if (startIndex < 0 || startIndex >= static_cast<int32_t>(GetWideText().length())) {
         return;
     }
     auto wideText = GetWideText().erase(startIndex, length);
